requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.9.0
blis==1.0.1
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.8
exceptiongroup==1.2.2
fastapi==0.115.5
filelock==3.16.1
Flask==3.1.0
fsspec==2024.10.0
grpcio==1.68.0
grpcio-tools==1.68.0
h11==0.14.0
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
langcodes==3.4.1
language_data==1.2.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.10
numpy==2.0.2
packaging==24.2
preshed==3.0.9
protobuf==5.28.3
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
shellingham==1.5.4
smart-open==7.0.5
sniffio==1.3.1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.41.2
thinc==8.3.2
tokenizers==0.20.3
tqdm==4.67.0
transformers==4.46.2
typer==0.13.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.16.0
zipp==3.21.0
